Criando pacotes com dominio invertido e movendo arquivos.
