+++ src/br/com/alura/bytebank/modelo/Auxiliar.kt
@@ -1,4 +1,4 @@
-package modelo
+package br.com.alura.bytebank.modelo
 
 class Auxiliar(
     nome: String,
