+++ src/br/com/alura/bytebank/modelo/CalculadoraBonificacao.kt
@@ -1,4 +1,4 @@
-package modelo
+package br.com.alura.bytebank.modelo
 
 class CalculadoraBonificacao {
