+++ src/br/com/alura/bytebank/modelo/Cliente.kt
@@ -1,4 +1,4 @@
-package modelo
+package br.com.alura.bytebank.modelo
 
 class Cliente(
     val nome: String,
