+++ src/br/com/alura/bytebank/modelo/Diretor.kt
@@ -1,4 +1,4 @@
-package modelo
+package br.com.alura.bytebank.modelo
 
 class Diretor(
     nome: String,
