+++ src/br/com/alura/bytebank/modelo/Gerente.kt
@@ -1,4 +1,4 @@
-package modelo
+package br.com.alura.bytebank.modelo
 
 class Gerente(
     nome: String,
