+++ src/br/com/alura/bytebank/novomodelo/Cliente.kt
@@ -1,0 +1,4 @@
+package br.com.alura.bytebank.novomodelo
+
+class Cliente {
+}
