+++ src/br/com/alura/bytebank/teste/testaAutenticacao.kt
@@ -1,9 +1,9 @@
-package teste
+package br.com.alura.bytebank.teste
 
-import modelo.Cliente
-import modelo.Diretor
-import modelo.Gerente
-import modelo.SistemaInterno
+import br.com.alura.bytebank.modelo.Cliente
+import br.com.alura.bytebank.modelo.Diretor
+import br.com.alura.bytebank.modelo.Gerente
+import br.com.alura.bytebank.modelo.SistemaInterno
 
 fun testaAutenticacao() {
     val gerente = Gerente(
