+++ src/br/com/alura/bytebank/teste/testaCondicoes.kt
@@ -1,4 +1,4 @@
-package teste
+package br.com.alura.bytebank.teste
 
 fun testaCondicoes(saldo: Double) {
     if (saldo > 0.0) {
