+++ src/br/com/alura/bytebank/teste/testaLacos.kt
@@ -1,4 +1,4 @@
-package teste
+package br.com.alura.bytebank.teste
 
 fun testaLacos() {
     var i = 0
